Split util
