+++ sig/big_decimal.rbs
@@ -1047,19 +1047,6 @@ class BigDecimal < Numeric
   #
   def zero?: () -> bool
 
-  # <!--
-  #   rdoc-file=lib/bigdecimal/util.rb
-  #   - a.to_d -> bigdecimal
-  # -->
-  # Returns self.
-  #
-  #     require 'bigdecimal/util'
-  #
-  #     d = BigDecimal("3.14")
-  #     d.to_d                       # => 0.314e1
-  #
-  def to_d: () -> BigDecimal
-
   private
 
   def initialize_copy: (self) -> self
@@ -1256,21 +1243,6 @@ end
 %a{annotate:rdoc:skip}
 class Integer
   # <!--
-  #   rdoc-file=lib/bigdecimal/util.rb
-  #   - int.to_d  -> bigdecimal
-  # -->
-  # Returns the value of `int` as a BigDecimal.
-  #
-  #     require 'bigdecimal'
-  #     require 'bigdecimal/util'
-  #
-  #     42.to_d   # => 0.42e2
-  #
-  # See also Kernel.BigDecimal.
-  #
-  def to_d: () -> BigDecimal
-
-  # <!--
   #   rdoc-file=numeric.c
   #   - self / numeric -> numeric_result
   # -->
@@ -1342,27 +1314,6 @@ end
 %a{annotate:rdoc:skip}
 class Float
   # <!--
-  #   rdoc-file=lib/bigdecimal/util.rb
-  #   - float.to_d             -> bigdecimal
-  #   - float.to_d(precision)  -> bigdecimal
-  # -->
-  # Returns the value of `float` as a BigDecimal. The `precision` parameter is
-  # used to determine the number of significant digits for the result. When
-  # `precision` is set to `0`, the number of digits to represent the float being
-  # converted is determined automatically. The default `precision` is `0`.
-  #
-  #     require 'bigdecimal'
-  #     require 'bigdecimal/util'
-  #
-  #     0.5.to_d         # => 0.5e0
-  #     1.234.to_d       # => 0.1234e1
-  #     1.234.to_d(2)    # => 0.12e1
-  #
-  # See also Kernel.BigDecimal.
-  #
-  def to_d: (?Integer precision) -> BigDecimal
-
-  # <!--
   #   rdoc-file=numeric.c
   #   - self / other -> numeric
   # -->
@@ -1424,48 +1375,7 @@ class Float
 end
 
 %a{annotate:rdoc:skip}
-class String
-  # <!--
-  #   rdoc-file=lib/bigdecimal/util.rb
-  #   - str.to_d  -> bigdecimal
-  # -->
-  # Returns the result of interpreting leading characters in `str` as a
-  # BigDecimal.
-  #
-  #     require 'bigdecimal'
-  #     require 'bigdecimal/util'
-  #
-  #     "0.5".to_d             # => 0.5e0
-  #     "123.45e1".to_d        # => 0.12345e4
-  #     "45.67 degrees".to_d   # => 0.4567e2
-  #
-  # See also Kernel.BigDecimal.
-  #
-  def to_d: () -> BigDecimal
-end
-
-%a{annotate:rdoc:skip}
 class Rational
-  # <!--
-  #   rdoc-file=lib/bigdecimal/util.rb
-  #   - rat.to_d(precision)  -> bigdecimal
-  # -->
-  # Returns the value as a BigDecimal.
-  #
-  # The `precision` parameter is used to determine the number of significant
-  # digits for the result. When `precision` is set to `0`, the number of digits to
-  # represent the float being converted is determined automatically. The default
-  # `precision` is `0`.
-  #
-  #     require 'bigdecimal'
-  #     require 'bigdecimal/util'
-  #
-  #     Rational(22, 7).to_d(3)   # => 0.314e1
-  #
-  # See also Kernel.BigDecimal.
-  #
-  def to_d: (Integer precision) -> BigDecimal
-
   # <!--
   #   rdoc-file=rational.c
   #   - rat / numeric     ->  numeric
@@ -1531,30 +1441,6 @@ end
 %a{annotate:rdoc:skip}
 class Complex
   # <!--
-  #   rdoc-file=lib/bigdecimal/util.rb
-  #   - cmp.to_d             -> bigdecimal
-  #   - cmp.to_d(precision)  -> bigdecimal
-  # -->
-  # Returns the value as a BigDecimal. If the imaginary part is not `0`, an error
-  # is raised
-  #
-  # The `precision` parameter is used to determine the number of significant
-  # digits for the result. When `precision` is set to `0`, the number of digits to
-  # represent the float being converted is determined automatically. The default
-  # `precision` is `0`.
-  #
-  #     require 'bigdecimal'
-  #     require 'bigdecimal/util'
-  #
-  #     Complex(0.1234567, 0).to_d(4)   # => 0.1235e0
-  #     Complex(Rational(22, 7), 0).to_d(3)   # => 0.314e1
-  #     Complex(1, 1).to_d   # raises ArgumentError
-  #
-  # See also Kernel.BigDecimal.
-  #
-  def to_d: (*untyped args) -> BigDecimal
-
-  # <!--
   #   rdoc-file=complex.c
   #   - complex / numeric -> new_complex
   # -->
@@ -1613,20 +1499,4 @@ class Complex
   #
   def -: (BigDecimal) -> Complex
        | ...
-end
-
-%a{annotate:rdoc:skip}
-class NilClass
-  # <!--
-  #   rdoc-file=lib/bigdecimal/util.rb
-  #   - nil.to_d -> bigdecimal
-  # -->
-  # Returns nil represented as a BigDecimal.
-  #
-  #     require 'bigdecimal'
-  #     require 'bigdecimal/util'
-  #
-  #     nil.to_d   # => 0.0
-  #
-  def to_d: () -> BigDecimal
 end
