+++ test_sig/test_big_decimal.rb
@@ -1,7 +1,7 @@
 require "bigdecimal"
-require "bigdecimal/util"
 require 'test/unit'
 require 'rbs/unit_test'
+require_relative './test_helper'
 
 class BigDecimalSingletonTest < Test::Unit::TestCase
   include TestHelper
@@ -495,11 +495,6 @@ class BigDecimalTest < Test::Unit::TestCase
     end
   end
 
-  def test_to_d
-    assert_send_type "() -> ::BigDecimal",
-                     BigDecimal("1.23"), :to_d
-  end
-
   def test_to_f
     assert_send_type "() -> ::Float",
                      BigDecimal("1.23"), :to_f
@@ -521,10 +516,6 @@ class IntegerToBigDecimalTest < Test::Unit::TestCase
 
   library "bigdecimal"
   testing "::Integer"
-
-  def test_to_d_with_integer
-    assert_send_type "() -> ::BigDecimal", 123, :to_d
-  end
 
   def test_plus_with_integer
     assert_send_type "(::BigDecimal) -> ::BigDecimal",
@@ -553,10 +544,6 @@ class FloatToBigDecimalTest < Test::Unit::TestCase
   library "bigdecimal"
   testing "::Float"
 
-  def test_to_d_with_float
-    assert_send_type "() -> ::BigDecimal", 12.3, :to_d
-  end
-
   def test_plus_with_float
     assert_send_type "(::BigDecimal) -> ::BigDecimal",
                      1.23, :+, BigDecimal("1.23")
@@ -578,26 +565,11 @@ class FloatToBigDecimalTest < Test::Unit::TestCase
   end
 end
 
-class StringToBigDecimalTest < Test::Unit::TestCase
-  include TestHelper
-
-  library "bigdecimal"
-  testing "::String"
-
-  def test_to_d_with_string
-    assert_send_type "() -> ::BigDecimal", "123", :to_d
-  end
-end
-
 class RationalToBigDecimalTest < Test::Unit::TestCase
   include TestHelper
 
   library "bigdecimal"
   testing "::Rational"
-
-  def test_to_d_with_rational
-    assert_send_type "(Integer) -> ::BigDecimal", Rational(22, 7), :to_d, 3
-  end
 
   def test_plus_with_rational
     assert_send_type "(::BigDecimal) -> ::BigDecimal",
@@ -626,10 +598,6 @@ class ComplexToBigDecimalTest < Test::Unit::TestCase
   library "bigdecimal"
   testing "::Complex"
 
-  def test_to_d_with_complex
-    assert_send_type "() -> ::BigDecimal", Complex(0.1234567, 0), :to_d
-  end
-
   def test_plus_with_complex
     assert_send_type "(::BigDecimal) -> ::Complex",
                      Complex(0.1234567, 0), :+, BigDecimal("1.23")
@@ -648,16 +616,5 @@ class ComplexToBigDecimalTest < Test::Unit::TestCase
   def test_multiply_with_complex
     assert_send_type "(::BigDecimal) -> ::Complex",
                      Complex(0.1234567, 0), :*, BigDecimal("1.23")
-  end
-end
-
-class NilToBigDecimalTest < Test::Unit::TestCase
-  include TestHelper
-
-  library "bigdecimal"
-  testing "::NilClass"
-
-  def test_to_d_with_nil
-    assert_send_type "() -> ::BigDecimal", nil, :to_d
   end
 end
